show eip notices on eip pages
