+++ src/components/eip/index.ts
@@ -1,3 +1,4 @@
 export { EipPage } from './EipPage';
 export { EipTimeline } from './EipTimeline';
 export { EipSearch } from './EipSearch';
+export { EipNotice } from './EipNotice';
