+++ src/components/network-upgrade/EipCard.tsx
@@ -14,6 +14,7 @@ import {
   getForkRelationship,
 } from '../../utils';
 import { Tooltip, CopyLinkButton } from '../ui';
+import { EipNotice } from '../eip/EipNotice';
 // Butterfly view disabled — data is stale. Uncomment to re-enable.
 // import { useButterflyData } from '../../hooks/useButterflyData';
 // import { ClientTestingProgress } from './ClientTestingProgress';
@@ -24,24 +25,6 @@ interface EipCardProps {
   handleExternalLinkClick: (linkType: string, url: string) => void;
   cardId?: string;
   showCopyLink?: boolean;
-}
-
-const NOTICE_CLASSES = {
-  container: 'bg-amber-50 dark:bg-amber-900/10 border-amber-200 dark:border-amber-700/40',
-  icon: 'text-amber-600 dark:text-amber-300',
-  title: 'text-amber-900 dark:text-amber-100',
-  text: 'text-amber-800 dark:text-amber-200',
-  link: 'text-amber-700 hover:text-amber-900 dark:text-amber-300 dark:hover:text-amber-100',
-};
-
-function formatNoticeCallReference(call: string, timestamp?: number): { display: string; link: string } {
-  const [prefix, number] = call.split('/');
-  const paddedNumber = number.padStart(3, '0');
-  const baseLink = `/calls/${prefix}/${paddedNumber}`;
-  return {
-    display: `${prefix.toUpperCase()} #${number}`,
-    link: timestamp ? `${baseLink}#t=${timestamp}` : baseLink,
-  };
 }
 
 export const EipCard: React.FC<EipCardProps> = ({
@@ -207,42 +190,7 @@ export const EipCard: React.FC<EipCardProps> = ({
 
       {/* Description */}
       <div className="">
-        {notice && (() => {
-          const callReference = notice.call
-            ? formatNoticeCallReference(notice.call, notice.timestamp)
-            : null;
-
-          return (
-            <div className={`mb-4 rounded border p-4 ${NOTICE_CLASSES.container}`}>
-              <div className="flex items-start gap-3">
-                <svg
-                  className={`mt-0.5 h-4 w-4 flex-shrink-0 ${NOTICE_CLASSES.icon}`}
-                  fill="none"
-                  stroke="currentColor"
-                  strokeWidth="2"
-                  viewBox="0 0 24 24"
-                  aria-hidden="true"
-                >
-                  <path strokeLinecap="round" strokeLinejoin="round" d="M12 9v4m0 4h.01M10.29 3.86 1.82 18a2 2 0 0 0 1.71 3h16.94a2 2 0 0 0 1.71-3L13.71 3.86a2 2 0 0 0-3.42 0Z" />
-                </svg>
-                <div className="min-w-0">
-                  <p className={`text-sm font-semibold ${NOTICE_CLASSES.title}`}>{notice.title}</p>
-                  <p className={`mt-1 text-sm leading-relaxed ${NOTICE_CLASSES.text}`}>
-                    {parseMarkdownLinks(notice.text)}
-                  </p>
-                  {callReference && (
-                    <Link
-                      to={callReference.link}
-                      className={`mt-2 inline-flex items-center gap-1 text-xs font-medium underline decoration-1 underline-offset-2 transition-colors ${NOTICE_CLASSES.link}`}
-                    >
-                      Source: {callReference.display}
-                    </Link>
-                  )}
-                </div>
-              </div>
-            </div>
-          );
-        })()}
+        {notice && <EipNotice notice={notice} className="mb-4" />}
 
         <p className="text-slate-700 dark:text-slate-300 text-sm leading-relaxed">
           {parseMarkdownLinks(getSummaryDescription(eip))}
